Añado movimientos y clase padre
