+++ Assets/Scripts/CharacterController.cs
@@ -1,0 +1,43 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public abstract class CharacterController : MonoBehaviour
+{
+
+    [Header("Parameters")]
+    protected Rigidbody _rb;
+    [SerializeField] float rigidBodySpeed = 0;
+    protected float _xAxis = 0;
+    protected float _yAxis = 0;
+    protected float _zAxis = 0;
+    protected Animator _anim;
+
+    [Header("Animation")]
+    [SerializeField] protected string _xAxisName = "xAxis";
+    [SerializeField] protected string _zAxisName = "zAxis";
+
+    private void Awake()
+    {
+        _rb = GetComponent<Rigidbody>();
+    }
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        _rb.constraints = RigidbodyConstraints.FreezeRotation;
+        _anim = GetComponentInChildren<Animator>();
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+    protected void movement()
+
+    {
+        Vector3 dir = (transform.right * _xAxis + transform.forward * _zAxis).normalized;
+        _rb.MovePosition(transform.position += dir * rigidBodySpeed * Time.fixedDeltaTime);
+    }
+}
