+++ Assets/Scripts/CharacterController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cb897cea2363ca54798f802006f692a4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
